Removed the material files
